+++ src/modules/booking-agent/langgraph/langgraph-graph.service.ts
@@ -1,129 +1,21 @@
-import { Annotation, END, START, StateGraph } from "@langchain/langgraph";
+import { END, START, StateGraph } from "@langchain/langgraph";
 import { Injectable, Logger } from "@nestjs/common";
-import { CarNotAvailableException } from "../../booking/booking.error";
-import { BookingCreationService } from "../../booking/booking-creation.service";
-import { DatabaseService } from "../../database/database.service";
-import { GooglePlacesService } from "../../maps/google-places.service";
-import { getMissingRequiredFields } from "../booking-agent.helper";
-import { BookingAgentSearchService } from "../booking-agent-search.service";
-import {
-  LANGGRAPH_NODE_NAMES,
-  LANGGRAPH_OUTBOUND_MODE,
-  LANGGRAPH_SERVICE_UNAVAILABLE_MESSAGE,
-} from "./langgraph.const";
+import { CreateBookingNode } from "./create-booking.node";
+import { ExtractNode } from "./extract.node";
+import { HandoffNode } from "./handoff.node";
+import { LANGGRAPH_NODE_NAMES } from "./langgraph.const";
 import { LangGraphExecutionFailedException } from "./langgraph.error";
 import type {
-  AgentResponse,
-  BookingAgentLocationValidationState,
   BookingAgentState,
-  BookingDraft,
-  BookingStage,
-  ConversationMessage,
-  ExtractionResult,
-  InteractiveReply,
   LangGraphInvokeInput,
   LangGraphInvokeResult,
-  LangGraphOutboxItem,
-  LocationValidationState,
-  UserPreferences,
-  VehicleSearchOption,
 } from "./langgraph.interface";
-import {
-  convertToExtractedParams,
-  createDefaultLocationValidationState,
-} from "./langgraph.interface";
-import { buildBookingInputFromDraft, buildGuestIdentity } from "./langgraph-booking-orchestrator";
-import {
-  applyDerivedDraftFields,
-  hasDraftChanged,
-  shouldApplyDraftPatch,
-} from "./langgraph-booking-rules";
-import { LangGraphExtractorService } from "./langgraph-extractor.service";
-import { buildOutboxItems } from "./langgraph-outbox.builder";
-import { LangGraphResponderService } from "./langgraph-responder.service";
-import { resolveRouteDecision } from "./langgraph-router.policy";
+import { AnnotationState, BookingAgentAnnotation } from "./langgraph-state.annotation";
 import { LangGraphStateService } from "./langgraph-state.service";
-
-/**
- * Helper to create an Annotation channel with a default value.
- * Required when you want a default but no special reducer logic.
- */
-function AnnotationWithDefault<T>(defaultValue: T) {
-  return Annotation<T>({
-    reducer: (_current: T, update: T) => update,
-    default: () => defaultValue,
-  });
-}
-
-const BookingAgentAnnotation = Annotation.Root({
-  messages: Annotation<ConversationMessage[]>({
-    reducer: (current, update) => [...current, ...update],
-    default: () => [],
-  }),
-  conversationId: Annotation<string>,
-  customerId: AnnotationWithDefault<string | null>(null),
-  inboundMessage: AnnotationWithDefault<string>(""),
-  inboundMessageId: AnnotationWithDefault<string>(""),
-  inboundInteractive: AnnotationWithDefault<InteractiveReply | undefined>(undefined),
-  draft: Annotation<BookingDraft & { __clear?: boolean }>({
-    reducer: (current, update) => {
-      // Special flag to clear the draft completely
-      if ("__clear" in update && update.__clear === true) {
-        const { __clear: _, ...rest } = update;
-        return rest;
-      }
-      return { ...current, ...update };
-    },
-    default: () => ({}),
-  }),
-  stage: Annotation<BookingStage>({
-    reducer: (_, update) => update,
-    default: () => "greeting",
-  }),
-  turnCount: Annotation<number>({
-    reducer: (current, update) => (typeof update === "number" ? update : current + 1),
-    default: () => 0,
-  }),
-  extraction: AnnotationWithDefault<ExtractionResult | null>(null),
-  availableOptions: Annotation<VehicleSearchOption[]>({
-    reducer: (_, update) => update,
-    default: () => [],
-  }),
-  lastShownOptions: Annotation<VehicleSearchOption[]>({
-    reducer: (_, update) => update,
-    default: () => [],
-  }),
-  selectedOption: AnnotationWithDefault<VehicleSearchOption | null>(null),
-  holdId: AnnotationWithDefault<string | null>(null),
-  holdExpiresAt: AnnotationWithDefault<string | null>(null),
-  bookingId: AnnotationWithDefault<string | null>(null),
-  paymentLink: AnnotationWithDefault<string | null>(null),
-  preferences: Annotation<UserPreferences & { __clear?: boolean }>({
-    reducer: (current, update) => {
-      // Special flag to clear preferences completely
-      if ("__clear" in update && update.__clear === true) {
-        const { __clear: _, ...rest } = update;
-        return rest;
-      }
-      return { ...current, ...update };
-    },
-    default: () => ({}),
-  }),
-  response: AnnotationWithDefault<AgentResponse | null>(null),
-  outboxItems: Annotation<LangGraphOutboxItem[]>({
-    reducer: (_, update) => update,
-    default: () => [],
-  }),
-  locationValidation: Annotation<BookingAgentState["locationValidation"]>({
-    reducer: (_, update) => update,
-    default: () => createDefaultLocationValidationState(),
-  }),
-  nextNode: AnnotationWithDefault<string | null>(null),
-  error: AnnotationWithDefault<string | null>(null),
-  statusMessage: AnnotationWithDefault<string | null>(null),
-});
-
-type AnnotationState = typeof BookingAgentAnnotation.State;
+import { MergeNode } from "./merge.node";
+import { RespondNode } from "./respond.node";
+import { RouteNode } from "./route.node";
+import { SearchNode } from "./search.node";
 
 @Injectable()
 export class LangGraphGraphService {
@@ -132,12 +24,13 @@ export class LangGraphGraphService {
 
   constructor(
     private readonly stateService: LangGraphStateService,
-    private readonly extractorService: LangGraphExtractorService,
-    private readonly responderService: LangGraphResponderService,
-    private readonly bookingAgentSearchService: BookingAgentSearchService,
-    private readonly bookingCreationService: BookingCreationService,
-    private readonly databaseService: DatabaseService,
-    private readonly googlePlacesService: GooglePlacesService,
+    private readonly extractNode: ExtractNode,
+    private readonly mergeNode: MergeNode,
+    private readonly routeNode: RouteNode,
+    private readonly searchNode: SearchNode,
+    private readonly createBookingNode: CreateBookingNode,
+    private readonly respondNode: RespondNode,
+    private readonly handoffNode: HandoffNode,
   ) {}
 
   async invoke(input: LangGraphInvokeInput): Promise<LangGraphInvokeResult> {
@@ -171,25 +64,19 @@ export class LangGraphGraphService {
       this.stateService.addMessage(initialState, "user", message);
 
       const graph = this.getOrBuildGraph();
-      const result = await graph.invoke(initialState, {
+      const finalStateResult = await graph.invoke(initialState, {
         configurable: { thread_id: conversationId },
       });
 
-      const finalState = result as BookingAgentState;
-
-      if (finalState.response) {
-        this.stateService.addMessage(finalState, "assistant", finalState.response.text);
+      if (finalStateResult.response) {
+        this.stateService.addMessage(finalStateResult, "assistant", finalStateResult.response.text);
       }
 
-      await this.stateService.saveState(conversationId, finalState);
+      await this.stateService.saveState(conversationId, finalStateResult);
 
-      return {
-        response: finalState.response,
-        outboxItems: finalState.outboxItems,
-        stage: finalState.stage,
-        draft: finalState.draft,
-        error: finalState.error,
-      };
+      const { response, outboxItems, stage, draft, error } = finalStateResult;
+
+      return { response, outboxItems, stage, draft, error };
     } catch (error) {
       this.logger.error("Graph execution failed", {
         conversationId,
@@ -212,13 +99,16 @@ export class LangGraphGraphService {
 
   private buildGraph() {
     const workflow = new StateGraph(BookingAgentAnnotation)
-      .addNode(LANGGRAPH_NODE_NAMES.EXTRACT, this.extractNode.bind(this))
-      .addNode(LANGGRAPH_NODE_NAMES.MERGE, this.mergeNode.bind(this))
-      .addNode(LANGGRAPH_NODE_NAMES.ROUTE, this.routeNode.bind(this))
-      .addNode(LANGGRAPH_NODE_NAMES.SEARCH, this.searchNode.bind(this))
-      .addNode(LANGGRAPH_NODE_NAMES.CREATE_BOOKING, this.createBookingNode.bind(this))
-      .addNode(LANGGRAPH_NODE_NAMES.RESPOND, this.respondNode.bind(this))
-      .addNode(LANGGRAPH_NODE_NAMES.HANDOFF, this.handoffNode.bind(this))
+      .addNode(LANGGRAPH_NODE_NAMES.EXTRACT, this.extractNode.run.bind(this.extractNode))
+      .addNode(LANGGRAPH_NODE_NAMES.MERGE, this.mergeNode.run.bind(this.mergeNode))
+      .addNode(LANGGRAPH_NODE_NAMES.ROUTE, this.routeNode.run.bind(this.routeNode))
+      .addNode(LANGGRAPH_NODE_NAMES.SEARCH, this.searchNode.run.bind(this.searchNode))
+      .addNode(
+        LANGGRAPH_NODE_NAMES.CREATE_BOOKING,
+        this.createBookingNode.run.bind(this.createBookingNode),
+      )
+      .addNode(LANGGRAPH_NODE_NAMES.RESPOND, this.respondNode.run.bind(this.respondNode))
+      .addNode(LANGGRAPH_NODE_NAMES.HANDOFF, this.handoffNode.run.bind(this.handoffNode))
       .addEdge(START, LANGGRAPH_NODE_NAMES.EXTRACT)
       .addEdge(LANGGRAPH_NODE_NAMES.EXTRACT, LANGGRAPH_NODE_NAMES.MERGE)
       .addEdge(LANGGRAPH_NODE_NAMES.MERGE, LANGGRAPH_NODE_NAMES.ROUTE)
@@ -231,974 +121,7 @@ export class LangGraphGraphService {
     return workflow.compile();
   }
 
-  private async extractNode(state: AnnotationState): Promise<Partial<AnnotationState>> {
-    try {
-      const extraction = await this.extractorService.extract(state);
-      this.logger.log("Extract node completed", {
-        intent: extraction.intent,
-        draftPatch: extraction.draftPatch,
-        confidence: extraction.confidence,
-      });
-      // Clear any prior system outage state once extraction succeeds.
-      return { extraction, error: null };
-    } catch (error) {
-      this.logger.error("Extract node failed", { error });
-      // When extraction fails (external service error), clear state and show user-friendly message.
-      // Don't expose raw errors like "429 quota exceeded" to users.
-      return {
-        extraction: {
-          intent: "unknown",
-          draftPatch: {},
-          confidence: 0,
-        },
-        error: LANGGRAPH_SERVICE_UNAVAILABLE_MESSAGE,
-        statusMessage: null,
-        draft: { __clear: true },
-        preferences: { __clear: true },
-        availableOptions: [],
-        lastShownOptions: [],
-        selectedOption: null,
-        locationValidation: createDefaultLocationValidationState(),
-        stage: "greeting",
-      };
-    }
-  }
-
-  private mergeNode(state: AnnotationState): Partial<AnnotationState> {
-    const { extraction, draft, preferences } = state;
-
-    if (!extraction) {
-      return {};
-    }
-
-    const baseDraft = shouldApplyDraftPatch(extraction.intent)
-      ? { ...draft, ...extraction.draftPatch }
-      : { ...draft };
-    const newDraft = applyDerivedDraftFields(baseDraft, state.inboundMessage);
-
-    const newPreferences = this.mergePreferencesWithHint(preferences, extraction.preferenceHint);
-
-    const draftChanged = hasDraftChanged(draft, newDraft);
-    const pickupLocationChanged = draft.pickupLocation !== newDraft.pickupLocation;
-    const dropoffLocationChanged = draft.dropoffLocation !== newDraft.dropoffLocation;
-    const shouldClearOptions = draftChanged && state.availableOptions.length > 0;
-
-    this.logger.debug("Merge node completed", {
-      newDraft,
-      autoFilledDropoffLocation: !draft.dropoffLocation && !!newDraft.dropoffLocation,
-      autoFilledDropoffDate: !draft.dropoffDate && !!newDraft.dropoffDate,
-    });
-
-    const nextLocationValidation = this.nextLocationValidationOnDraftMerge(
-      state.locationValidation,
-      pickupLocationChanged,
-      dropoffLocationChanged,
-    );
-
-    return {
-      draft: newDraft,
-      preferences: newPreferences,
-      availableOptions: shouldClearOptions ? [] : state.availableOptions,
-      lastShownOptions: shouldClearOptions ? [] : state.lastShownOptions,
-      locationValidation: nextLocationValidation,
-    };
-  }
-
-  private mergePreferencesWithHint(
-    preferences: BookingAgentState["preferences"],
-    preferenceHint: string | undefined,
-  ): BookingAgentState["preferences"] {
-    const newPreferences = { ...preferences };
-    if (!preferenceHint) {
-      return newPreferences;
-    }
-
-    if (preferenceHint === "cheaper" || preferenceHint === "budget") {
-      newPreferences.pricePreference = "budget";
-    } else if (preferenceHint === "premium" || preferenceHint === "luxury") {
-      newPreferences.pricePreference = "premium";
-    }
-
-    const existingNotes = newPreferences.notes ?? [];
-    newPreferences.notes = existingNotes.includes(preferenceHint)
-      ? existingNotes
-      : [...existingNotes, preferenceHint];
-    return newPreferences;
-  }
-
-  private getLocationValidationState(
-    state: Pick<BookingAgentState, "locationValidation">,
-  ): BookingAgentLocationValidationState {
-    return state.locationValidation ?? createDefaultLocationValidationState();
-  }
-
-  private nextLocationValidationOnDraftMerge(
-    current: BookingAgentState["locationValidation"],
-    pickupLocationChanged: boolean,
-    dropoffLocationChanged: boolean,
-  ): BookingAgentLocationValidationState {
-    const previous = current ?? createDefaultLocationValidationState();
-    return {
-      pickup: pickupLocationChanged
-        ? {
-            status: "unvalidated",
-            lastValidatedInput: null,
-            normalizedAddress: null,
-          }
-        : previous.pickup,
-      dropoff: dropoffLocationChanged
-        ? {
-            status: "unvalidated",
-            lastValidatedInput: null,
-            normalizedAddress: null,
-          }
-        : previous.dropoff,
-    };
-  }
-
-  private shouldValidateLocationField(
-    locationValue: string | undefined,
-    validation: LocationValidationState,
-  ): boolean {
-    const normalizedInput = locationValue?.trim();
-    if (!normalizedInput) {
-      return false;
-    }
-
-    if (validation.lastValidatedInput !== normalizedInput) {
-      return true;
-    }
-
-    return validation.status === "unvalidated";
-  }
-
-  private shouldBlockSearchForInvalidLocation(
-    locationValue: string | undefined,
-    validation: LocationValidationState,
-  ): boolean {
-    if (!locationValue?.trim()) {
-      return false;
-    }
-
-    return validation.status === "invalid";
-  }
-
-  private buildLocationFailureStatusMessage(
-    locationValue: string | undefined,
-    locationLabel: "pickup" | "drop-off",
-  ): string {
-    const value = locationValue?.trim() || "that location";
-    return this.buildFullAddressPrompt(value, locationLabel);
-  }
-
-  private resolveInvalidLocationStatus(): LocationValidationState["status"] {
-    return "invalid";
-  }
-
-  private routeNode(state: AnnotationState): Partial<AnnotationState> {
-    const { extraction, draft, stage, availableOptions, selectedOption } = state;
-
-    if (state.error === LANGGRAPH_SERVICE_UNAVAILABLE_MESSAGE) {
-      return {
-        nextNode: LANGGRAPH_NODE_NAMES.RESPOND,
-        stage: "greeting",
-      };
-    }
-
-    const missingFields = getMissingRequiredFields(draft);
-    this.logger.log("Route node decision", {
-      intent: extraction?.intent,
-      stage,
-      missingFields,
-      hasSelectedOption: !!selectedOption,
-      availableOptionsCount: availableOptions.length,
-      draft: {
-        bookingType: draft.bookingType,
-        pickupDate: draft.pickupDate,
-        pickupTime: draft.pickupTime,
-        dropoffDate: draft.dropoffDate,
-        pickupLocation: draft.pickupLocation,
-        dropoffLocation: draft.dropoffLocation,
-      },
-    });
-
-    const decision = resolveRouteDecision(state as BookingAgentState);
-    const isControlIntent =
-      extraction?.intent === "new_booking" ||
-      extraction?.intent === "reset" ||
-      extraction?.intent === "greeting" ||
-      extraction?.intent === "cancel" ||
-      extraction?.intent === "request_agent";
-    const shouldRunEarlyPickupValidation =
-      !isControlIntent &&
-      (decision.nextNode ?? LANGGRAPH_NODE_NAMES.RESPOND) === LANGGRAPH_NODE_NAMES.RESPOND &&
-      (decision.stage ?? stage) === "collecting" &&
-      !!draft.pickupLocation &&
-      this.shouldValidateLocationField(
-        draft.pickupLocation,
-        this.getLocationValidationState(state).pickup,
-      );
-    if (shouldRunEarlyPickupValidation) {
-      return {
-        ...decision,
-        nextNode: LANGGRAPH_NODE_NAMES.SEARCH,
-        stage: "collecting",
-      };
-    }
-
-    return decision;
-  }
-
   private routeDecision(state: AnnotationState): string {
     return state.nextNode ?? LANGGRAPH_NODE_NAMES.RESPOND;
-  }
-
-  private async searchNode(state: AnnotationState): Promise<Partial<AnnotationState>> {
-    try {
-      const validationResult = await this.validateAndNormalizeLocations(state);
-      if (validationResult.earlyReturn) {
-        return {
-          ...validationResult.earlyReturn,
-          draft: validationResult.draft,
-          locationValidation: validationResult.locationValidation,
-        };
-      }
-      const validatedDraft = validationResult.draft;
-      const locationValidation = validationResult.locationValidation;
-
-      if (
-        this.shouldBlockSearchForInvalidLocation(
-          validatedDraft.pickupLocation,
-          locationValidation.pickup,
-        )
-      ) {
-        return {
-          draft: validatedDraft,
-          stage: "collecting",
-          availableOptions: [],
-          lastShownOptions: [],
-          statusMessage: this.buildLocationFailureStatusMessage(
-            validatedDraft.pickupLocation,
-            "pickup",
-          ),
-          locationValidation,
-        };
-      }
-
-      if (
-        this.shouldBlockSearchForInvalidLocation(
-          validatedDraft.dropoffLocation,
-          locationValidation.dropoff,
-        )
-      ) {
-        return {
-          draft: validatedDraft,
-          stage: "collecting",
-          availableOptions: [],
-          lastShownOptions: [],
-          statusMessage: this.buildLocationFailureStatusMessage(
-            validatedDraft.dropoffLocation,
-            "drop-off",
-          ),
-          locationValidation,
-        };
-      }
-
-      const missingFields = getMissingRequiredFields(validatedDraft);
-      if (missingFields.length > 0) {
-        return {
-          draft: validatedDraft,
-          stage: "collecting",
-          availableOptions: [],
-          lastShownOptions: [],
-          error: null,
-          locationValidation,
-        };
-      }
-
-      const extractedParams = convertToExtractedParams(validatedDraft);
-      this.logger.log(
-        {
-          draft: state.draft,
-          extractedParams,
-        },
-        "Search node executing",
-      );
-      const searchResult = await this.bookingAgentSearchService.searchVehiclesFromExtracted(
-        extractedParams,
-        "",
-      );
-
-      // Check if there's a precondition issue (missing/invalid field)
-      if (searchResult.precondition) {
-        this.logger.warn("Search returned precondition", {
-          precondition: searchResult.precondition,
-          extractedParams,
-        });
-        // Go back to collecting stage to ask for the missing field
-        // Preserve the validated draft and location lookup state
-        return {
-          draft: validatedDraft,
-          availableOptions: [],
-          lastShownOptions: [],
-          stage: "collecting",
-          error: null,
-          statusMessage: searchResult.precondition.prompt,
-          locationValidation,
-        };
-      }
-
-      const hasStrictVehicleFilters = this.hasStrictVehicleFilters(validatedDraft);
-      const allowAlternativeMatches = this.shouldAllowAlternativeMatches(state.preferences);
-      const shouldSuppressAlternatives =
-        searchResult.exactMatches.length === 0 &&
-        searchResult.alternatives.length > 0 &&
-        hasStrictVehicleFilters &&
-        !allowAlternativeMatches;
-
-      const options: VehicleSearchOption[] = shouldSuppressAlternatives
-        ? []
-        : [...searchResult.exactMatches, ...searchResult.alternatives].slice(0, 5);
-
-      const newStage: BookingStage = options.length > 0 ? "presenting_options" : "collecting";
-
-      this.logger.log("Search node completed", {
-        exactMatchCount: searchResult.exactMatches.length,
-        alternativeCount: searchResult.alternatives.length,
-        totalOptions: options.length,
-        newStage,
-        optionDetails: options.map((o) => ({
-          id: o.id,
-          make: o.make,
-          model: o.model,
-          price: o.estimatedTotalInclVat,
-        })),
-      });
-
-      // If no vehicles found, set a status message so the responder can inform the user
-      const noResultsMessage = this.buildSearchStatusMessage({
-        optionsCount: options.length,
-        exactMatchCount: searchResult.exactMatches.length,
-        alternativeCount: searchResult.alternatives.length,
-        shouldSuppressAlternatives,
-        draft: validatedDraft,
-      });
-
-      return {
-        draft: validatedDraft,
-        availableOptions: options,
-        lastShownOptions: options,
-        stage: newStage,
-        error: null,
-        statusMessage: noResultsMessage,
-        locationValidation,
-      };
-    } catch (error) {
-      this.logger.error("Search node failed", {
-        error: error instanceof Error ? error.message : String(error),
-        stack: error instanceof Error ? error.stack : undefined,
-      });
-      return {
-        stage: "collecting",
-        error: LANGGRAPH_SERVICE_UNAVAILABLE_MESSAGE,
-        availableOptions: [],
-        lastShownOptions: [],
-        statusMessage:
-          "I couldn't complete the car search right now. Please try again in a moment, and I'll search again.",
-      };
-    }
-  }
-
-  private shouldAllowAlternativeMatches(preferences: BookingAgentState["preferences"]): boolean {
-    return Boolean(preferences.notes?.includes("show_alternatives"));
-  }
-
-  private hasStrictVehicleFilters(draft: BookingDraft): boolean {
-    return Boolean(
-      draft.color?.trim() ||
-        draft.vehicleType ||
-        draft.serviceTier ||
-        draft.make?.trim() ||
-        draft.model?.trim(),
-    );
-  }
-
-  private buildSearchStatusMessage(input: {
-    optionsCount: number;
-    exactMatchCount: number;
-    alternativeCount: number;
-    shouldSuppressAlternatives: boolean;
-    draft: BookingDraft;
-  }): string | null {
-    const { optionsCount, exactMatchCount, alternativeCount, shouldSuppressAlternatives, draft } =
-      input;
-    if (shouldSuppressAlternatives && alternativeCount > 0) {
-      return `I couldn't find an exact match for ${this.buildVehicleConstraintLabel(draft)}. I found close alternatives instead - would you like me to show them?`;
-    }
-
-    if (optionsCount === 0) {
-      return "No vehicles matching your criteria are available for the selected date. Would you like to try a different date, vehicle type, or booking type?";
-    }
-
-    if (exactMatchCount === 0) {
-      return "I couldn't find an exact match, but I found close alternatives you can choose from.";
-    }
-
-    return null;
-  }
-
-  private buildVehicleConstraintLabel(draft: BookingDraft): string {
-    const parts = [
-      draft.color?.trim(),
-      draft.vehicleType?.toLowerCase().replaceAll("_", " "),
-      draft.serviceTier?.toLowerCase().replaceAll("_", " "),
-    ]
-      .filter(Boolean)
-      .map(String);
-    if (parts.length > 0) {
-      return `a ${parts.join(" ")}`;
-    }
-
-    const makeModel = [draft.make?.trim(), draft.model?.trim()].filter(Boolean).join(" ");
-    return makeModel ? `"${makeModel}"` : "your requested vehicle";
-  }
-
-  private async respondNode(state: AnnotationState): Promise<Partial<AnnotationState>> {
-    try {
-      if (state.outboxItems.length > 0 && state.response) {
-        return {};
-      }
-
-      this.logger.log(
-        {
-          stage: state.stage,
-          availableOptionsCount: state.availableOptions.length,
-          lastShownOptionsCount: state.lastShownOptions.length,
-          hasSelectedOption: !!state.selectedOption,
-        },
-        "Respond node executing",
-      );
-      this.logger.debug("Respond node draft details", { draft: state.draft });
-
-      const response = await this.responderService.generateResponse(state);
-
-      const outboxItems = buildOutboxItems(state, response);
-
-      return { response, outboxItems };
-    } catch (error) {
-      this.logger.error("Respond node failed", { error });
-      return {
-        response: {
-          text: "I'm having trouble right now. Please try again or type AGENT to speak with someone.",
-        },
-        error: error instanceof Error ? error.message : String(error),
-      };
-    }
-  }
-
-  private buildFullAddressPrompt(
-    locationValue: string,
-    locationLabel: "pickup" | "drop-off" = "pickup",
-  ): string {
-    const locationNoun = locationLabel === "pickup" ? "pickup address" : "drop-off address";
-    return [
-      `I couldn't confirm "${locationValue}" as a complete ${locationNoun}.`,
-      "",
-      `Please share the full ${locationNoun} (for example: building number + street, or a hotel/landmark name in Lagos).`,
-    ].join("\n");
-  }
-
-  private async validateAndNormalizeLocations(state: AnnotationState): Promise<{
-    draft: BookingDraft;
-    locationValidation: BookingAgentLocationValidationState;
-    earlyReturn?: Partial<AnnotationState>;
-  }> {
-    const pickupResult = await this.validateAndNormalizePickupLocation(state);
-    const dropoffResult = await this.validateAndNormalizeDropoffLocation(
-      state,
-      pickupResult.draft,
-      pickupResult.locationValidation,
-    );
-
-    // Validate both fields in the same turn, but prioritize pickup prompt when both are invalid.
-    const earlyReturn = pickupResult.earlyReturn ?? dropoffResult.earlyReturn;
-
-    return {
-      draft: dropoffResult.draft,
-      locationValidation: dropoffResult.locationValidation,
-      earlyReturn,
-    };
-  }
-
-  private async validateAndNormalizePickupLocation(state: AnnotationState): Promise<{
-    draft: BookingDraft;
-    locationValidation: BookingAgentLocationValidationState;
-    earlyReturn?: Partial<AnnotationState>;
-  }> {
-    const locationValidation = this.getLocationValidationState(state);
-    const pickupValidation = locationValidation.pickup;
-    const shouldValidate = this.shouldValidateLocationField(
-      state.draft.pickupLocation,
-      pickupValidation,
-    );
-
-    if (!shouldValidate) {
-      return { draft: state.draft, locationValidation };
-    }
-
-    const pickupInput = state.draft.pickupLocation?.trim() ?? "";
-    const locationResult = await this.googlePlacesService.validateAddress(pickupInput);
-
-    if (!locationResult.isValid) {
-      const nextLocationValidation: BookingAgentLocationValidationState = {
-        ...locationValidation,
-        pickup: {
-          status: this.resolveInvalidLocationStatus(),
-          lastValidatedInput: pickupInput,
-          normalizedAddress: null,
-        },
-      };
-      return {
-        draft: state.draft,
-        locationValidation: nextLocationValidation,
-        earlyReturn: this.buildInvalidPickupLocationResult(
-          state,
-          pickupInput,
-          nextLocationValidation,
-        ),
-      };
-    }
-
-    if (!locationResult.normalizedAddress) {
-      return { draft: state.draft, locationValidation };
-    }
-
-    const normalizedAddress = locationResult.normalizedAddress;
-    const wasDropoffSameAsPickup = state.draft.dropoffLocation === state.draft.pickupLocation;
-    const nextLocationValidation: BookingAgentLocationValidationState = {
-      ...locationValidation,
-      pickup: {
-        status: "valid",
-        lastValidatedInput: normalizedAddress,
-        normalizedAddress,
-      },
-      ...(wasDropoffSameAsPickup && {
-        dropoff: {
-          status: "valid",
-          lastValidatedInput: normalizedAddress,
-          normalizedAddress,
-        },
-      }),
-    };
-
-    return {
-      draft: {
-        ...state.draft,
-        pickupLocation: normalizedAddress,
-        // Keep dropoff in sync if it was auto-filled from pickup via "same location"
-        ...(wasDropoffSameAsPickup && {
-          dropoffLocation: normalizedAddress,
-        }),
-      },
-      locationValidation: nextLocationValidation,
-    };
-  }
-
-  private async validateAndNormalizeDropoffLocation(
-    state: AnnotationState,
-    draft: BookingDraft,
-    locationValidation: BookingAgentLocationValidationState,
-  ): Promise<{
-    draft: BookingDraft;
-    locationValidation: BookingAgentLocationValidationState;
-    earlyReturn?: Partial<AnnotationState>;
-  }> {
-    const pickupLocation = draft.pickupLocation?.trim();
-    const dropoffLocation = draft.dropoffLocation?.trim();
-
-    if (!pickupLocation || !dropoffLocation) {
-      return { draft, locationValidation };
-    }
-
-    if (pickupLocation === dropoffLocation) {
-      if (locationValidation.pickup.status === "valid") {
-        const normalizedAddress = locationValidation.pickup.normalizedAddress ?? pickupLocation;
-        return {
-          draft,
-          locationValidation: {
-            ...locationValidation,
-            dropoff: {
-              status: "valid",
-              lastValidatedInput: normalizedAddress,
-              normalizedAddress,
-            },
-          },
-        };
-      }
-
-      if (
-        locationValidation.pickup.status === "invalid" &&
-        locationValidation.pickup.lastValidatedInput === pickupLocation
-      ) {
-        return {
-          draft,
-          locationValidation: {
-            ...locationValidation,
-            dropoff: {
-              status: "invalid",
-              lastValidatedInput: dropoffLocation,
-              normalizedAddress: null,
-            },
-          },
-        };
-      }
-    }
-
-    if (!this.shouldValidateLocationField(dropoffLocation, locationValidation.dropoff)) {
-      return { draft, locationValidation };
-    }
-
-    const locationResult = await this.googlePlacesService.validateAddress(dropoffLocation);
-    if (!locationResult.isValid) {
-      const nextLocationValidation: BookingAgentLocationValidationState = {
-        ...locationValidation,
-        dropoff: {
-          status: this.resolveInvalidLocationStatus(),
-          lastValidatedInput: dropoffLocation,
-          normalizedAddress: null,
-        },
-      };
-      return {
-        draft,
-        locationValidation: nextLocationValidation,
-        earlyReturn: this.buildInvalidDropoffLocationResult(
-          state,
-          dropoffLocation,
-          nextLocationValidation,
-        ),
-      };
-    }
-
-    if (!locationResult.normalizedAddress) {
-      return { draft, locationValidation };
-    }
-
-    const normalizedAddress = locationResult.normalizedAddress;
-    return {
-      draft: {
-        ...draft,
-        dropoffLocation: normalizedAddress,
-      },
-      locationValidation: {
-        ...locationValidation,
-        dropoff: {
-          status: "valid",
-          lastValidatedInput: normalizedAddress,
-          normalizedAddress,
-        },
-      },
-    };
-  }
-
-  private buildInvalidPickupLocationResult(
-    state: AnnotationState,
-    pickupLocation: string,
-    locationValidation: BookingAgentLocationValidationState,
-  ): Partial<AnnotationState> {
-    const suggestionText = this.buildFullAddressPrompt(pickupLocation || "that location", "pickup");
-
-    return {
-      stage: "collecting",
-      locationValidation: {
-        ...locationValidation,
-        pickup: {
-          status: this.resolveInvalidLocationStatus(),
-          lastValidatedInput: pickupLocation,
-          normalizedAddress: null,
-        },
-      },
-      response: { text: suggestionText },
-      error: null,
-      outboxItems: [
-        {
-          conversationId: state.conversationId,
-          dedupeKey: `langgraph:${state.inboundMessageId}:address-checking`,
-          mode: LANGGRAPH_OUTBOUND_MODE.FREE_FORM,
-          textBody: "Thanks - I'm checking that pickup address on Google Maps now...",
-        },
-        {
-          conversationId: state.conversationId,
-          dedupeKey: `langgraph:${state.inboundMessageId}:address-suggestions`,
-          mode: LANGGRAPH_OUTBOUND_MODE.FREE_FORM,
-          textBody: suggestionText,
-        },
-      ],
-    };
-  }
-
-  private buildInvalidDropoffLocationResult(
-    state: AnnotationState,
-    dropoffLocation: string,
-    locationValidation: BookingAgentLocationValidationState,
-  ): Partial<AnnotationState> {
-    const suggestionText = this.buildFullAddressPrompt(dropoffLocation, "drop-off");
-
-    return {
-      stage: "collecting",
-      locationValidation: {
-        ...locationValidation,
-        dropoff: {
-          status: this.resolveInvalidLocationStatus(),
-          lastValidatedInput: dropoffLocation,
-          normalizedAddress: null,
-        },
-      },
-      response: { text: suggestionText },
-      error: null,
-      outboxItems: [
-        {
-          conversationId: state.conversationId,
-          dedupeKey: `langgraph:${state.inboundMessageId}:dropoff-address-checking`,
-          mode: LANGGRAPH_OUTBOUND_MODE.FREE_FORM,
-          textBody: "Thanks - I'm checking that drop-off address on Google Maps now...",
-        },
-        {
-          conversationId: state.conversationId,
-          dedupeKey: `langgraph:${state.inboundMessageId}:dropoff-address-suggestions`,
-          mode: LANGGRAPH_OUTBOUND_MODE.FREE_FORM,
-          textBody: suggestionText,
-        },
-      ],
-    };
-  }
-
-  private async createBookingNode(state: AnnotationState): Promise<Partial<AnnotationState>> {
-    const { draft, selectedOption } = state;
-
-    if (!selectedOption) {
-      this.logger.error("Create booking node called without selected option");
-      return {
-        error: "No vehicle selected for booking",
-        stage: "confirming",
-      };
-    }
-
-    try {
-      const conversation = await this.getConversationForBooking(state.conversationId);
-      if (!conversation) {
-        return {
-          error: "Unable to create booking - conversation not found",
-          stage: "confirming",
-        };
-      }
-
-      const validationError = this.validateDraftBeforeBookingCreation(draft);
-      if (validationError) {
-        return validationError;
-      }
-
-      this.logger.log("Creating booking", {
-        conversationId: state.conversationId,
-        phone: conversation.phoneE164,
-        vehicleId: selectedOption.id,
-        draft,
-      });
-
-      const guestIdentity = buildGuestIdentity(conversation.phoneE164, conversation.profileName);
-      const {
-        input: bookingInput,
-        normalizedStartDate,
-        normalizedEndDate,
-      } = buildBookingInputFromDraft(draft, selectedOption, guestIdentity);
-
-      this.logBookingCreationInput(bookingInput, normalizedStartDate, normalizedEndDate);
-
-      // For WhatsApp bookings, we create as a guest booking
-      const result = await this.bookingCreationService.createBooking(
-        bookingInput,
-        null, // No session user for WhatsApp bookings
-      );
-
-      this.logger.log("Booking created successfully", {
-        bookingId: result.bookingId,
-        checkoutUrl: result.checkoutUrl,
-      });
-
-      return {
-        bookingId: result.bookingId,
-        paymentLink: result.checkoutUrl,
-        stage: "awaiting_payment",
-      };
-    } catch (error) {
-      this.logBookingCreationFailure(state, selectedOption, error);
-
-      if (error instanceof CarNotAvailableException) {
-        const fallbackOptions = await this.fetchFreshOptionsForDraft(
-          state.draft,
-          selectedOption.id,
-        );
-        if (fallbackOptions.length > 0) {
-          return {
-            selectedOption: null,
-            availableOptions: fallbackOptions,
-            lastShownOptions: fallbackOptions,
-            stage: "presenting_options",
-            statusMessage:
-              "That vehicle is no longer available for your selected date and time. Here are some alternatives.",
-          };
-        }
-
-        return {
-          selectedOption: null,
-          availableOptions: [],
-          lastShownOptions: [],
-          stage: "collecting",
-          error: null,
-          statusMessage:
-            "That vehicle is no longer available for your selected date and time. Please adjust your date, booking type, or vehicle preference.",
-        };
-      }
-
-      // Keep a user-safe fallback message for unexpected booking failures.
-      return {
-        error: LANGGRAPH_SERVICE_UNAVAILABLE_MESSAGE,
-        statusMessage: null,
-        stage: "confirming",
-      };
-    }
-  }
-
-  private async getConversationForBooking(conversationId: string) {
-    const conversation = await this.databaseService.whatsAppConversation.findUnique({
-      where: { id: conversationId },
-      select: {
-        phoneE164: true,
-        profileName: true,
-      },
-    });
-
-    if (!conversation) {
-      this.logger.error("Conversation not found for booking creation", { conversationId });
-    }
-
-    return conversation;
-  }
-
-  private validateDraftBeforeBookingCreation(draft: BookingDraft): Partial<AnnotationState> | null {
-    if (!draft.pickupDate || !draft.dropoffDate || !draft.pickupTime) {
-      this.logger.error(
-        { draft },
-        "Missing required date/time fields in draft - cannot create booking",
-      );
-      return {
-        error:
-          "Missing required booking details. Please provide pickup date, drop-off date, and pickup time.",
-        stage: "collecting",
-      };
-    }
-
-    return null;
-  }
-
-  private async fetchFreshOptionsForDraft(
-    draft: BookingDraft,
-    excludedOptionId?: string,
-  ): Promise<VehicleSearchOption[]> {
-    try {
-      const extractedParams = convertToExtractedParams(draft);
-      const searchResult = await this.bookingAgentSearchService.searchVehiclesFromExtracted(
-        extractedParams,
-        "",
-      );
-
-      if (searchResult.precondition) {
-        return [];
-      }
-
-      const options = [...searchResult.exactMatches, ...searchResult.alternatives]
-        .filter((option) => option.id !== excludedOptionId)
-        .slice(0, 5);
-
-      this.logger.log("Fetched fresh options after booking unavailability", {
-        excludedOptionId,
-        optionCount: options.length,
-      });
-
-      return options;
-    } catch (fallbackError) {
-      this.logger.warn("Failed to fetch fresh options after booking unavailability", {
-        error: fallbackError instanceof Error ? fallbackError.message : String(fallbackError),
-      });
-      return [];
-    }
-  }
-
-  private logBookingCreationInput(
-    bookingInput: {
-      carId: string;
-      pickupAddress: string;
-      bookingType: string;
-      pickupTime?: string;
-      clientTotalAmount?: string;
-      sameLocation?: boolean;
-      guestEmail?: string;
-    },
-    normalizedStartDate: Date,
-    normalizedEndDate: Date,
-  ): void {
-    this.logger.log(
-      {
-        carId: bookingInput.carId,
-        startDate: normalizedStartDate.toISOString(),
-        endDate: normalizedEndDate.toISOString(),
-        pickupAddress: bookingInput.pickupAddress,
-        bookingType: bookingInput.bookingType,
-        pickupTime: bookingInput.pickupTime,
-        clientTotalAmount: bookingInput.clientTotalAmount,
-        sameLocation: bookingInput.sameLocation,
-        guestEmail: bookingInput.guestEmail,
-      },
-      "Calling BookingCreationService.createBooking",
-    );
-  }
-
-  private logBookingCreationFailure(
-    state: AnnotationState,
-    selectedOption: BookingAgentState["selectedOption"],
-    error: unknown,
-  ): void {
-    const errorMessage = error instanceof Error ? error.message : String(error);
-    const errorName = error instanceof Error ? error.name : "UnknownError";
-
-    this.logger.error(
-      {
-        errorName,
-        errorMessage,
-        conversationId: state.conversationId,
-        draft: state.draft,
-        selectedOptionId: selectedOption?.id,
-        selectedOptionPrice: selectedOption?.estimatedTotalInclVat,
-        stack: error instanceof Error ? error.stack : undefined,
-      },
-      "Booking creation failed",
-    );
-  }
-
-  private handoffNode(state: AnnotationState): Partial<AnnotationState> {
-    return {
-      response: {
-        text: "A Tripdly agent will join this chat shortly. Please share your booking reference if available.",
-      },
-      outboxItems: [
-        {
-          conversationId: state.conversationId,
-          dedupeKey: `langgraph:handoff:${state.inboundMessageId}`,
-          mode: LANGGRAPH_OUTBOUND_MODE.FREE_FORM,
-          textBody:
-            "A Tripdly agent will join this chat shortly. Please share your booking reference if available.",
-        },
-      ],
-      stage: "cancelled",
-    };
   }
 }
