+++ src/modules/booking-agent/langgraph/langgraph-node-state.interface.ts
@@ -1,0 +1,8 @@
+import type { BookingAgentState, BookingDraft, UserPreferences } from "./langgraph.interface";
+
+export type LangGraphNodeState = BookingAgentState;
+
+export type LangGraphNodeResult = Partial<BookingAgentState> & {
+  draft?: BookingDraft & { __clear?: boolean };
+  preferences?: UserPreferences & { __clear?: boolean };
+};
